backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
openai==0.28.0
djangorestframework-simplejwt==5.3.1
google-auth==2.35.0
